pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "crispy-bootstrap5>=2024.10",
    "django>=5.1.6",
    "django-crispy-forms>=2.3",
    "django-import-export>=4.3.5",
    "django-simple-history>=3.8.0",
    "django-unfold>=0.47.0",
    "isort>=6.0.0",
    "lxml>=5.3.0",
    "requests>=2.32.3",
]
